+++ lib/cassandra/cluster.rb
@@ -286,18 +286,13 @@ module Cassandra
       promise = @futures.promise
 
       @control_connection.close_async.on_complete do |f|
-        # The control connection stops the reactor as part of closing, but stop
-        # it here as well so that a closed cluster can never leave its reactor
-        # thread behind, whatever state the control connection was in.
-        @io_reactor.stop.on_complete do |_|
-          if f.resolved?
-            promise.fulfill(self)
-          else
-            f.on_failure {|e| promise.break(e)}
-          end
-
-          @executor.shutdown
+        if f.resolved?
+          promise.fulfill(self)
+        else
+          f.on_failure {|e| promise.break(e)}
         end
+
+        @executor.shutdown
       end
 
       promise.future